The Reign of Chaos! This is a lot of hacks, which destroy my code just for compilation and correct work on win10.

+++ src/formfactors.cpp
@@ -5,13 +5,14 @@
 #include "formfactors.h"
 
 #include <algorithm>
+#include <ctime>
 #include <iostream>
 #include <random>
 
 #include <embree3/rtcore.h>
 
 
-std::vector<glm::vec2> GenerateRandomSamples(const uint samplesNumber) {
+std::vector<glm::vec2> GenerateRandomSamples(const unsigned samplesNumber) {
     std::default_random_engine generator;
     std::uniform_real_distribution<float> distribution(0, 1);
     std::vector<glm::vec2> samples(samplesNumber);
@@ -68,7 +69,7 @@ float GetInfluenceForTwoSamples(
         return 0;
     }
     bool hasIntersection = false;
-    for (uint interQuadIdx = 0; interQuadIdx < quadsToCheck.size() && !hasIntersection; ++interQuadIdx) {
+    for (unsigned interQuadIdx = 0; interQuadIdx < quadsToCheck.size() && !hasIntersection; ++interQuadIdx) {
         hasIntersection = quadsToCheck[interQuadIdx].CheckIntersectionWithVector(sample1, sample2);
     }
     if (hasIntersection) {
@@ -83,7 +84,7 @@ inline float ComputeFormFactorForTwoQuads(
     const Quad& q2,
     const std::vector<Quad>& quadsToCheck
 ) {
-    static const uint samplesNum = 20;
+    static const unsigned samplesNum = 20;
     static const auto samples = GenerateRandomSamples(samplesNum);
 
     float result = 0;
@@ -93,13 +94,13 @@ inline float ComputeFormFactorForTwoQuads(
     std::array<int, th_num> cnt_ar{};
     res_ar.fill(0);
     cnt_ar.fill(0);
-    for (uint samplerIdx1 = 0; samplerIdx1 < samplesNum; ++samplerIdx1) {
+    for (unsigned samplerIdx1 = 0; samplerIdx1 < samplesNum; ++samplerIdx1) {
         const glm::vec4 sample1 = q1.GetSample(samples[samplerIdx1]);
 #pragma omp parallel for num_threads(th_num)
-        for (uint th_i = 0; th_i < th_num; ++th_i) {
-            const auto begin = static_cast<uint>(samples.size() / th_num * th_i);
-            const auto end = static_cast<uint>(std::min(samples.size() / th_num * (th_i + 1), samples.size()));
-            for (uint samplerIdx2 = begin; samplerIdx2 < end; ++samplerIdx2) {
+        for (unsigned th_i = 0; th_i < th_num; ++th_i) {
+            const auto begin = static_cast<unsigned>(samples.size() / th_num * th_i);
+            const auto end = static_cast<unsigned>(std::min(samples.size() / th_num * (th_i + 1), samples.size()));
+            for (unsigned samplerIdx2 = begin; samplerIdx2 < end; ++samplerIdx2) {
                 const glm::vec4 sample2 = q2.GetSample(samples[samplerIdx2]);
                 const float sampleValue = GetInfluenceForTwoSamples(
                     sample1,
@@ -116,7 +117,7 @@ inline float ComputeFormFactorForTwoQuads(
             }
         }
     }
-    for (uint th_i = 0; th_i < th_num; ++th_i) {
+    for (unsigned th_i = 0; th_i < th_num; ++th_i) {
         result += res_ar[th_i];
         count += cnt_ar[th_i];
     }
@@ -128,7 +129,7 @@ inline float ComputeFormFactorForTwoQuads(
 }
 
 
-std::vector<Quad> FilterQuadsForChecking(const uint quadIdx1, const uint quadIdx2, const std::vector<Quad>& quads) {
+std::vector<Quad> FilterQuadsForChecking(const unsigned quadIdx1, const unsigned quadIdx2, const std::vector<Quad>& quads) {
     const glm::vec4 vecBegin = quads[quadIdx1].GetSample(glm::vec2(0.5, 0.5));
     const glm::vec4 line4 = quads[quadIdx2].GetSample(glm::vec2(0.5, 0.5)) - vecBegin;
     const glm::vec3 line = {line4.x, line4.y, line4.z};
@@ -136,7 +137,7 @@ std::vector<Quad> FilterQuadsForChecking(const uint quadIdx1, const uint quadIdx
     const float thresholdDist = quads[0].GetMaxSide() * std::sqrt(2.f);
 
     std::vector<Quad> result;
-    for (uint i = 0; i < quads.size(); ++i) {
+    for (unsigned i = 0; i < quads.size(); ++i) {
         if (i == quadIdx1 || i == quadIdx2) {
             continue;
         }
@@ -161,11 +162,11 @@ std::vector<std::vector<float> > ComputeFormFactors(
     for (auto &formFactor : formFactors) {
         formFactor.assign(quads.size(), 0);
     }
-    for (uint quadIdx1 = 0; quadIdx1 < quads.size(); ++quadIdx1) {
+    for (unsigned quadIdx1 = 0; quadIdx1 < quads.size(); ++quadIdx1) {
         std::vector<glm::vec4> samplesFirst;
-        uint beginQuad2 = antiradiance ? 0 : quadIdx1 + 1;
+        unsigned beginQuad2 = antiradiance ? 0 : quadIdx1 + 1;
         auto quad1 = quads[quadIdx1];
-        for (uint quadIdx2 = beginQuad2; quadIdx2 < quads.size(); ++quadIdx2) {
+        for (unsigned quadIdx2 = beginQuad2; quadIdx2 < quads.size(); ++quadIdx2) {
             if (quadIdx1 == quadIdx2) {
                 continue;
             }
@@ -189,7 +190,7 @@ std::vector<std::vector<float> > ComputeFormFactors(
             }
         }
         if (verbose) {
-            const ulong size = quads.size();
+            const long long size = quads.size();
             if (quadIdx1 * 100 / size < (quadIdx1 + 1) * 100 / size) {
                 std::cout << "FF process "
                           << (quadIdx1 + 1) * 100 / size
@@ -222,17 +223,17 @@ class EmbreeFFJob {
     RTCDevice Device;
     RTCScene Scene;
     const std::vector<Quad>& Quads;
-    std::vector<std::map<uint, float> > FF;
+    std::vector<std::map<unsigned, float> > FF;
     RTCIntersectContext IntersectionContext;
 public:
     EmbreeFFJob(
         const std::vector<Quad>& quads,
         const std::vector<std::vector<glm::vec4> >& points,
-        const std::vector<std::vector<uint> >& indices
+        const std::vector<std::vector<unsigned> >& indices
     ): Quads(quads) {
         Device = rtcNewDevice(""); CHECK_EMBREE
         Scene = rtcNewScene(Device); CHECK_EMBREE
-        for (uint i = 0; i < points.size(); ++i) {
+        for (unsigned i = 0; i < points.size(); ++i) {
             RTCGeometry geometry = rtcNewGeometry(Device, RTC_GEOMETRY_TYPE_TRIANGLE); CHECK_EMBREE
             RTCBuffer indicesBuffer = rtcNewSharedBuffer(
                 Device,
@@ -263,12 +264,12 @@ public:
         rtcReleaseDevice(Device);
     }
 
-    std::vector<std::map<uint, float> > Execute() {
+    std::vector<std::map<unsigned, float> > Execute() {
         FF.resize(Quads.size());
-        const uint PACKET_SIZE = 16;
+        const unsigned PACKET_SIZE = 16;
         const auto samples = GenerateRandomSamples(16);
-        for (uint i = 0; i < FF.size(); ++i) {
-            for (uint j = i + 1; j < FF.size(); ++j) {
+        for (unsigned i = 0; i < FF.size(); ++i) {
+            for (unsigned j = i + 1; j < FF.size(); ++j) {
                 std::vector<std::pair<glm::vec4, glm::vec4> > rays;
                 for (const auto firstQuadSample : samples) {
                     const auto sample1 = Quads[i].GetSample(firstQuadSample);
@@ -276,12 +277,12 @@ public:
                         rays.emplace_back(std::make_pair(sample1, Quads[j].GetSample(secondQuadSample) - sample1));
                     }
                 }
-                uint visibilityCount = 0;
+                unsigned visibilityCount = 0;
                 float samplesSum = 0;
                 const float BIAS = 1e-5f;
                 std::vector<RTCRay16> raysPackets(rays.size() / PACKET_SIZE);
-                for (uint k = 0; k < rays.size(); k += PACKET_SIZE) {
-                    for (uint l = 0; l < PACKET_SIZE; ++l) {
+                for (unsigned k = 0; k < rays.size(); k += PACKET_SIZE) {
+                    for (unsigned l = 0; l < PACKET_SIZE; ++l) {
                         raysPackets[k / PACKET_SIZE].org_x[l] = rays[k + l].first.x;
                         raysPackets[k / PACKET_SIZE].org_y[l] = rays[k + l].first.y;
                         raysPackets[k / PACKET_SIZE].org_z[l] = rays[k + l].first.z;
@@ -296,13 +297,13 @@ public:
                     }
                 }
 
-                for (uint k = 0; k < rays.size(); k += PACKET_SIZE) {
+                for (unsigned k = 0; k < rays.size(); k += PACKET_SIZE) {
                     const int validMask = ~0u;
                     rtcOccluded16(&validMask, Scene, &IntersectionContext, &raysPackets[k / PACKET_SIZE]); CHECK_EMBREE
                 }
 
-                for (uint k = 0; k < rays.size(); k += PACKET_SIZE) {
-                    for (uint l = 0; l < PACKET_SIZE; ++l) {
+                for (unsigned k = 0; k < rays.size(); k += PACKET_SIZE) {
+                    for (unsigned l = 0; l < PACKET_SIZE; ++l) {
                         if (std::isinf(raysPackets[k / PACKET_SIZE].tfar[l])) {
                             continue;
                         }
@@ -328,10 +329,10 @@ public:
     }
 };
 
-std::vector<std::map<uint, float> > ComputeFormFactorsEmbree(
+std::vector<std::map<unsigned, float> > ComputeFormFactorsEmbree(
     const std::vector<Quad>& quads,
     const std::vector<std::vector<glm::vec4> >& points,
-    const std::vector<std::vector<uint> >& indices
+    const std::vector<std::vector<unsigned> >& indices
 ) {
     EmbreeFFJob job(quads, points, indices);
     return job.Execute();
@@ -355,7 +356,7 @@ class EmbreeHierarchicalFFJob {
             return 0;
         }
 
-        const uint PACKET_SIZE = 16;
+        const unsigned PACKET_SIZE = 16;
         const auto samples = GenerateRandomSamples(PACKET_SIZE * 2);
 
         std::vector<std::pair<glm::vec4, glm::vec4> > rays;
@@ -365,12 +366,12 @@ class EmbreeHierarchicalFFJob {
                 rays.emplace_back(std::make_pair(sample1, Quads.GetQuad(idx2).GetSample(secondQuadSample) - sample1));
             }
         }
-        uint visibilityCount = 0;
+        unsigned visibilityCount = 0;
         float samplesSum = 0;
         const float BIAS = 1e-7f;
         std::vector<RTCRay16> raysPackets(rays.size() / PACKET_SIZE);
-        for (uint k = 0; k < rays.size(); k += PACKET_SIZE) {
-            for (uint l = 0; l < PACKET_SIZE; ++l) {
+        for (unsigned k = 0; k < rays.size(); k += PACKET_SIZE) {
+            for (unsigned l = 0; l < PACKET_SIZE; ++l) {
                 raysPackets[k / PACKET_SIZE].org_x[l] = rays[k + l].first.x;
                 raysPackets[k / PACKET_SIZE].org_y[l] = rays[k + l].first.y;
                 raysPackets[k / PACKET_SIZE].org_z[l] = rays[k + l].first.z;
@@ -386,7 +387,7 @@ class EmbreeHierarchicalFFJob {
         }
 
 #pragma omp parallel for
-        for (uint k = 0; k < rays.size(); k += PACKET_SIZE) {
+        for (unsigned k = 0; k < rays.size(); k += PACKET_SIZE) {
             const int validMask = ~0u;
             rtcInitIntersectContext(&IntersectionContext); CHECK_EMBREE
             IntersectionContext.flags = RTC_INTERSECT_CONTEXT_FLAG_COHERENT;
@@ -394,8 +395,8 @@ class EmbreeHierarchicalFFJob {
             rtcOccluded16(&validMask, Scene, &IntersectionContext, &raysPackets[k / PACKET_SIZE]); CHECK_EMBREE
         }
 
-        for (uint k = 0; k < rays.size(); k += PACKET_SIZE) {
-            for (uint l = 0; l < PACKET_SIZE; ++l) {
+        for (unsigned k = 0; k < rays.size(); k += PACKET_SIZE) {
+            for (unsigned l = 0; l < PACKET_SIZE; ++l) {
                 if (raysPackets[k / PACKET_SIZE].tfar[l] < 0.f) {
                     continue;
                 }
@@ -429,18 +430,18 @@ public:
     EmbreeHierarchicalFFJob(
         QuadsContainer& quads,
         const std::vector<std::vector<glm::vec4> >& points,
-        const std::vector<std::vector<uint> >& indices
+        const std::vector<std::vector<unsigned> >& indices
     ): Quads(quads) {
         Device = rtcNewDevice(""); CHECK_EMBREE
         Scene = rtcNewScene(Device); CHECK_EMBREE
-        for (uint i = 0; i < points.size(); ++i) {
+        for (unsigned i = 0; i < points.size(); ++i) {
             RTCGeometry geometry = rtcNewGeometry(Device, RTC_GEOMETRY_TYPE_TRIANGLE); CHECK_EMBREE
-            uint* ptrIdx = reinterpret_cast<uint*>(rtcSetNewGeometryBuffer(geometry, RTC_BUFFER_TYPE_INDEX, 0, RTC_FORMAT_UINT3, sizeof(uint) * 3, indices[i].size() / 3)); CHECK_EMBREE
-            for (uint j = 0; j < indices[i].size(); ++j) {
+            unsigned* ptrIdx = reinterpret_cast<unsigned*>(rtcSetNewGeometryBuffer(geometry, RTC_BUFFER_TYPE_INDEX, 0, RTC_FORMAT_UINT3, sizeof(unsigned) * 3, indices[i].size() / 3)); CHECK_EMBREE
+            for (unsigned j = 0; j < indices[i].size(); ++j) {
                 ptrIdx[j] = indices[i][j];
             }
             glm::vec3* ptrPnt = reinterpret_cast<glm::vec3*>(rtcSetNewGeometryBuffer(geometry, RTC_BUFFER_TYPE_VERTEX, 0, RTC_FORMAT_FLOAT3, sizeof(glm::vec3), points[i].size())); CHECK_EMBREE
-            for (uint j = 0; j < points[i].size(); ++j) {
+            for (unsigned j = 0; j < points[i].size(); ++j) {
                 ptrPnt[j] =  glm::vec3(points[i][j]);
             }
             rtcCommitGeometry(geometry); CHECK_EMBREE
@@ -458,7 +459,7 @@ public:
         rtcReleaseDevice(Device);
     }
 
-    std::vector<std::map<int, float>> Execute(const uint maxDepth) {
+    std::vector<std::map<int, float>> Execute(const unsigned maxDepth) {
 //        std::vector<Quad> result;
 //        for (int i = 0; i < Quads.GetSize(); ++i) {
 //            bigQuads.push_back(Quads.GetQuad(i));
@@ -473,7 +474,7 @@ public:
 //            Quads.AddQuad(i);
 //        }
         FF.resize(static_cast<unsigned long>(Quads.GetSize()));
-        uint sum = 0;
+        unsigned sum = 0;
         for (int i = 0; i < Quads.GetSize(); ++i) {
             for (int j = i + 1; j < Quads.GetSize(); ++j) {
                 ProcessTwoQuads(i, j);
@@ -488,8 +489,8 @@ public:
 std::vector<std::map<int, float>> ComputeFormFactorsEmbree(
     QuadsContainer& quads,
     const std::vector<std::vector<glm::vec4> >& points,
-    const std::vector<std::vector<uint> >& indices,
-    const uint maxDepth
+    const std::vector<std::vector<unsigned> >& indices,
+    const unsigned maxDepth
 ) {
     EmbreeHierarchicalFFJob job(quads, points, indices);
     return job.Execute(maxDepth);
